created a controllers folder to contain all the crud routes.
